+++ HISTORY.md
@@ -3,6 +3,7 @@
 # 1.2
  - Use docJGenerator 1.6.10
  - Use groovy 4.0.26 (only uses for tests, the binaries have no dependencies
+ - Move the Viewport a ViewBox classes in the org.girod.javafx.svgimage package
  - Fix #47: Add a parameter for the Color of the svg background in the JavaFX to SVG parameters
  - Fix #48: Add a parameter to generate the viewbox in the JavaFX to SVG parameters
  - Fix #49: Add a parameter to not allow the transform on the root of the svg document in the JavaFX to SVG parameters
@@ -10,6 +11,7 @@
  - Fix #45: Mask not working
  - Fix #37: SVGImage's scaleTo doesn't seem to override svg width height parameters
  - Fix #50: The default LineCap, LineJoin, and miterlimit on Shapes are incorrect
+ - Fix #36: Allow to scale on parent element
 
 ## 1.1
  - Fix #32: Fix some cases where the resulting paths are null